Fix int to int64

+++ senzingrestservice/senzingrestservice.go
@@ -45,7 +45,7 @@ type SenzingRestServiceImpl struct {
 	Port                           int
 	SenzingEngineConfigurationJson string
 	SenzingModuleName              string
-	SenzingVerboseLogging          int
+	SenzingVerboseLogging          int64
 	UrlRoutePrefix                 string
 }
